Added creating transactions, inserting the items into the transactions_items and creating transaction_history

+++ database/migrations/2022_07_11_002253_create_transaction_items_table.php
@@ -17,7 +17,7 @@ class CreateTransactionItemsTable extends Migration
             $table->id();
             $table->unsignedBigInteger('transaction_id');
             $table->unsignedBigInteger('item_id');
-            $table->integer('quantity');
+            $table->double('quantity',11,2);
             $table->string('remarks')->nullable();
             $table->timestamp('updated_at');
         });
